+++ src/constants.js
@@ -1,0 +1,5 @@
+export const ITEM_TYPE_LENGTH = 19;
+export const ITEM_TYPE_SYMBOL = '0';
+
+export const ITEM_DESCRIPTION_LENGTH = 29;
+export const ITEM_DESCRIPTION_SYMBOL = ' ';